+++ modules/hanna-sprinkles/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. -->
+- feat: Upgrade `sitewideAlerts` to support renamed property `renderedAlert`
 - feat: Make `loadSprinkles`' "base inport URL" resolution more resilient
 
 ## 0.8.44 – 0.8.45